requirements.txt
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_core==0.3.72
langchain_huggingface==0.3.1
protobuf==6.31.1
python-dotenv==1.1.1
